feat: add Open UI definitions
