+++ packages/web-components/fast-components/package.json
@@ -22,7 +22,7 @@
     "clean:dist": "node ../../../build/clean.js dist",
     "doc": "api-extractor run --local",
     "doc:ci": "api-extractor run",
-    "build": "npm run generate-default-palettes && tsc -p ./tsconfig.json && rollup -c && npm run doc",
+    "build": "npm run generate-default-palettes && tsc -p ./tsconfig.json && yarn run generate-open-ui-definition && rollup -c && npm run doc",
     "dev": "tsc -p ./tsconfig.json -w",
     "tdd": "npm run dev & npm run test-chrome:watch",
     "prepare": "yarn clean:dist && yarn build",
@@ -33,7 +33,8 @@
     "start": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "generate-default-palettes": "node -r esm build/generate-default-palettes.js",
-    "test": "yarn test-chrome:verbose && yarn doc:ci",
+    "generate-open-ui-definition": "tsc -p ./tsconfig.definitions.json && node -r esm build/generate-open-ui-definition",
+    "test": "yarn run build && yarn test-chrome:verbose && yarn doc:ci",
     "test-node": "mocha --reporter min --exit dist/esm/__test__/setup-node.js './dist/esm/**/*.spec.js'",
     "test-node:watch": "mocha --reporter min --exit dist/esm/__test__/setup-node.js --watch './dist/esm/**/*.spec.js'",
     "test-node:verbose": "mocha --reporter spec --exit dist/esm/__test__/setup-node.js './dist/esm/**/*.spec.js'",
@@ -58,6 +59,7 @@
     "@types/karma": "^5.0.0",
     "@types/mocha": "^7.0.2",
     "@types/webpack-env": "^1.15.2",
+    "ajv": "^6.12.2",
     "chai": "^4.2.0",
     "chai-spies": "^1.0.0",
     "circular-dependency-plugin": "^5.2.0",
