+++ packages/web-components/fast-components/src/accordion/accordion-item.definition.ts
