+++ packages/web-components/fast-components/src/accordion/accordion.definition.ts
