+++ packages/web-components/fast-components/src/anchor/index.ts
@@ -1,7 +1,7 @@
 import { attr, customElement } from "@microsoft/fast-element";
 import { Anchor, AnchorTemplate as template } from "@microsoft/fast-foundation";
-import { AnchorStyles as styles } from "./anchor.styles";
 import { ButtonAppearance } from "../button";
+import { AnchorStyles as styles } from "./anchor.styles";
 
 /**
  * Types of anchor appearance.
