+++ packages/web-components/fast-components/src/anchored-region/anchored-region.stories.ts
@@ -2,8 +2,8 @@ import { STORY_RENDERED } from "@storybook/core-events";
 import addons from "@storybook/addons";
 import { Direction, RtlScrollConverter } from "@microsoft/fast-web-utilities";
 import { FASTDesignSystemProvider } from "../design-system-provider";
-import AnchoreRegionTemplate from "./fixtures/base.html";
 import { FASTAnchoredRegion } from "../anchored-region";
+import AnchoreRegionTemplate from "./fixtures/base.html";
 
 // Prevent tree-shaking
 FASTAnchoredRegion;
