+++ packages/web-components/fast-components/src/badge/badge.definition.ts
