+++ packages/web-components/fast-components/src/card/card.definition.ts
