+++ packages/web-components/fast-components/src/component-definitions.ts
@@ -1,27 +1,27 @@
-export * from "./accordion/fast-accordion-item.definition";
-export * from "./accordion/fast-accordion.definition";
-export * from "./anchor/fast-anchor.definition";
-export * from "./badge/fast-badge.definition";
-export * from "./button/fast-button.definition";
-export * from "./card/fast-card.definition";
-export * from "./checkbox/fast-checkbox.definition";
-export * from "./design-system-provider/fast-design-system-provider.definition";
-export * from "./dialog/fast-dialog.definition";
-export * from "./divider/fast-divider.definition";
-export * from "./flipper/fast-flipper.definition";
-export * from "./menu/fast-menu.definition";
-export * from "./menu-item/fast-menu-item.definition";
-export * from "./progress/fast-progress-ring.definition";
-export * from "./progress/fast-progress.definition";
-export * from "./radio/fast-radio.definition";
-export * from "./radio-group/fast-radio-group.definition";
-export * from "./slider/fast-slider.definition";
-export * from "./slider-label/fast-slider-label.definition";
-export * from "./switch/fast-switch.definition";
-export * from "./tabs/fast-tab-panel.definition";
-export * from "./tabs/fast-tab.definition";
-export * from "./tabs/fast-tabs.definition";
-export * from "./text-area/fast-text-area.definition";
-export * from "./text-field/fast-text-field.definition";
-export * from "./tree-item/fast-tree-item.definition";
-export * from "./tree-view/fast-tree-view.definition";
+export * from "./accordion/accordion-item.definition";
+export * from "./accordion/accordion.definition";
+export * from "./anchor/anchor.definition";
+export * from "./badge/badge.definition";
+export * from "./button/button.definition";
+export * from "./card/card.definition";
+export * from "./checkbox/checkbox.definition";
+export * from "./design-system-provider/design-system-provider.definition";
+export * from "./dialog/dialog.definition";
+export * from "./divider/divider.definition";
+export * from "./flipper/flipper.definition";
+export * from "./menu/menu.definition";
+export * from "./menu-item/menu-item.definition";
+export * from "./progress/progress-ring.definition";
+export * from "./progress/progress.definition";
+export * from "./radio/radio.definition";
+export * from "./radio-group/radio-group.definition";
+export * from "./slider/slider.definition";
+export * from "./slider-label/slider-label.definition";
+export * from "./switch/switch.definition";
+export * from "./tabs/tab-panel.definition";
+export * from "./tabs/tab.definition";
+export * from "./tabs/tabs.definition";
+export * from "./text-area/text-area.definition";
+export * from "./text-field/text-field.definition";
+export * from "./tree-item/tree-item.definition";
+export * from "./tree-view/tree-view.definition";
