+++ packages/web-components/fast-components/src/design-system-provider/design-system-provider.definition.ts
