+++ packages/web-components/fast-components/src/design-system-provider/design-system-provider.open-ui.definition.json
@@ -1,0 +1,4 @@
+{
+    "name": "Design system provider",
+    "url": "https://fast.design/docs/components/design-system-provider"
+}
