+++ packages/web-components/fast-components/src/menu-item/menu-item.definition.ts
@@ -1,6 +1,5 @@
 import { WebComponentDefinition } from "@microsoft/fast-tooling/dist/data-utilities/web-component";
 import { DataType } from "@microsoft/fast-tooling";
-import { MenuItemRole } from "@microsoft/fast-foundation";
 
 export const fastMenuItemDefinition: WebComponentDefinition = {
     version: 1,
@@ -27,11 +26,11 @@ export const fastMenuItemDefinition: WebComponentDefinition = {
                     name: "role",
                     type: DataType.boolean,
                     description: "The role attribute",
-                    default: MenuItemRole.menuitem,
+                    default: "menuitem",
                     values: [
-                        { name: MenuItemRole.menuitem },
-                        { name: MenuItemRole.menuitemcheckbox },
-                        { name: MenuItemRole.menuitemradio },
+                        { name: "menuitem" },
+                        { name: "menuitemcheckbox" },
+                        { name: "menuitemradio" },
                     ],
                     required: false,
                 },
