+++ packages/web-components/fast-components/src/progress/progress-ring.definition.ts
