+++ packages/web-components/fast-components/src/progress/progress.definition.ts
