+++ packages/web-components/fast-components/src/slider-label/slider-label.definition.ts
