+++ packages/web-components/fast-components/src/slider/slider.definition.ts
