+++ packages/web-components/fast-components/src/switch/switch.definition.ts
