+++ packages/web-components/fast-components/src/tabs/tab-panel.definition.ts
