+++ packages/web-components/fast-components/src/tabs/tab.definition.ts
