+++ packages/web-components/fast-components/src/tabs/tabs.definition.ts
