+++ packages/web-components/fast-components/src/tree-view/index.ts
@@ -1,5 +1,5 @@
 import { customElement } from "@microsoft/fast-element";
-import { TreeView, TreeViewTemplate as template } from "@microsoft/fast-foundation";
+import { TreeViewTemplate as template, TreeView } from "@microsoft/fast-foundation";
 import { TreeViewStyles as styles } from "./tree-view.styles";
 
 /**
