+++ packages/web-components/fast-components/tsconfig.definitions.json
@@ -1,0 +1,7 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "outDir": "temp"
+    },
+    "include": ["src/**/*.definition.ts"]
+}
